requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
APScheduler==3.9.1
argos-translate-files==1.1.4
argostranslate==1.9.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachelib==0.10.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==2.1.1
click==8.1.7
commonmark==0.9.1
ctranslate2==3.20.0
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.0
expiringdict==1.2.2
filelock==3.13.1
Flask==2.2.2
flask-babel==3.1.0
Flask-Limiter==2.6.3
Flask-Session==0.4.0
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
frozenlist==1.4.1
fsspec==2023.12.2
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
gspread==5.12.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langdetect==1.0.9
LexiLang==1.0.1
libretranslate==1.5.2
limits==3.7.0
llama-index==0.9.22
lxml==4.9.4
MarkupSafe==2.1.3
marshmallow==3.20.1
Morfessor==2.0.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.6.1
outcome==1.3.0.post0
packaging==23.1
pandas==2.1.4
polib==1.1.1
prometheus-client==0.15.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.3.4
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
sacremoses==0.0.53
selenium==4.16.0
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
stanza==1.1.1
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
torch==2.0.1
tqdm==4.66.1
translatehtml==1.5.2
trio==0.23.2
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.18
waitress==2.1.2
Werkzeug==2.2.2
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
